move preprocessDoc to utils/prepare_data
